@@ -4,74 +4,76 @@
   "private": true,
     "scripts": {
     "ng": "ng",
-    "build": "npm run write:env -s && ng build --prod",
+    "build": "npm run write:env -s && ng build",
     "start": "npm run write:env -s && ng serve --proxy-config proxy.conf.js",
     "serve:sw": "npm run build -s && npx http-server ./dist -p 4200",
-    "lint": "ng lint && stylelint \"src/**/*.scss\" --syntax scss && htmlhint \"src\" --config .htmlhintrc",
+    "lint": "ng lint && stylelint \"src/**/*.scss\" --syntax scss",
     "test": "npm run write:env -s && ng test",
     "test:ci": "npm run write:env -s && npm run lint -s && ng test --configuration=ci",
-    "e2e": "npm run write:env -s && ng e2e",
     "translations:extract": "ngx-translate-extract --input ./src --output ./src/translations/template.json --format=json --clean --sort",
     "docs": "hads ./docs -o",
     "write:env": "ngx-scripts env npm_package_version",
-    "prettier": "prettier --write \"./{src,e2e}/**/*.{ts,js,html,scss}\"",
-    "prettier:check": "prettier --list-different \"./{src,e2e}/**/*.{ts,js,html,scss}\"",
-    "postinstall": "npm run prettier -s",
+    "prettier": "prettier --write \"./src/**/*.{ts,js,html,scss}\"",
+    "prettier:check": "prettier --list-different \"./src/**/*.{ts,js,html,scss}\"",
+    "postinstall": "npm run prettier -s && husky install",
     "generate": "ng generate"
   },
   "dependencies": {
-    "@angular/animations": "^11.0.8",
-    "@angular/common": "^11.0.8",
-    "@angular/compiler": "^11.0.8",
-    "@angular/core": "^11.0.8",
-    "@angular/forms": "^11.0.8",
-    "@angular/localize": "^11.0.8",
-    "@angular/platform-browser": "^11.0.8",
-    "@angular/platform-browser-dynamic": "^11.0.8",
-    "@angular/router": "^11.0.8",
+    "@angular/animations": "^12.1.1",
+    "@angular/common": "^12.1.1",
+    "@angular/compiler": "^12.1.1",
+    "@angular/core": "^12.1.1",
+    "@angular/forms": "^12.1.1",
+    "@angular/localize": "^12.1.1",
+    "@angular/platform-browser": "^12.1.1",
+    "@angular/platform-browser-dynamic": "^12.1.1",
+    "@angular/router": "^12.1.1",
     "@ngx-translate/core": "^13.0.0",
-    "@angular/service-worker": "^11.0.8",
-    "@ng-bootstrap/ng-bootstrap": "^8.0.0",
-    "bootstrap": "^4.5.3",
+    "@angular/service-worker": "^12.1.1",
+    "@ng-bootstrap/ng-bootstrap": "^11.0.0-beta.2",
+    "bootstrap": "^5.0.2",
     "@fortawesome/fontawesome-free": "^5.15.1",
     "rxjs": "^6.6.3",
-    "tslib": "^2.0.1",
-    "zone.js": "~0.10.3"
+    "tslib": "^2.2.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-builders/jest": "^11.0.0",
-    "@angular/cli": "~11.0.6",
-    "@angular/compiler-cli": "^11.0.8",
-    "@angular/language-service": "^11.0.8",
-    "@angular-devkit/build-angular": "^0.1100.6",
+    "@angular-builders/jest": "^12.1.0",
+    "@angular-devkit/build-angular": "~12.1.1",
+    "@angular-eslint/builder": "~12.3.1",
+    "@angular-eslint/eslint-plugin": "~12.3.1",
+    "@angular-eslint/eslint-plugin-template": "~12.3.1",
+    "@angular-eslint/schematics": "~12.3.1",
+    "@angular-eslint/template-parser": "~12.3.1",
+    "@angular/cli": "~12.1.1",
+    "@angular/compiler-cli": "~12.1.1",
+    "@angular/language-service": "~12.1.1",
     "@biesbjerg/ngx-translate-extract": "^7.0.3",
     "@biesbjerg/ngx-translate-extract-marker": "^1.0.0",
-    "@ngx-rocket/scripts": "^5.0.0",
-    "@ngneat/until-destroy": "^8.0.3",
-    "@types/jasmine": "^3.6.2",
-    "@types/jasminewd2": "^2.0.8",
+    "@ngx-rocket/scripts": "^5.1.0",
+    "@ngneat/until-destroy": "^8.0.4",
+    "@typescript-eslint/eslint-plugin": "4.28.2",
+    "@typescript-eslint/parser": "4.28.2",
     "@types/jest": "^26.0.19",
     "@types/node": "^12.19.12",
-    "codelyzer": "^6.0.1",
+    "eslint": "^7.26.0",
+    "eslint-plugin-import": "latest",
+    "eslint-plugin-jsdoc": "latest",
+    "eslint-plugin-prefer-arrow": "latest",
     "hads": "^3.0.0",
-    "htmlhint": "^0.14.2",
+    "htmlhint": "^0.15.1",
     "https-proxy-agent": "^5.0.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~6.0.0",
-    "jest": "^26.6.3",
+    "jest": "^27.0.6",
     "prettier": "^2.2.1",
-    "tslint-config-prettier": "^1.18.0",
     "stylelint-config-prettier": "^8.0.2",
     "pretty-quick": "^3.1.0",
-    "husky": "^4.3.6",
-    "protractor": "~7.0.0",
-    "stylelint": "~13.8.0",
-    "stylelint-config-recommended-scss": "~4.2.0",
-    "stylelint-config-standard": "~20.0.0",
-    "stylelint-scss": "~3.18.0",
-    "ts-node": "^9.1.1",
-    "tslint": "~6.1.0",
-    "typescript": "~4.0.5"
+    "husky": "^7.0.1",
+    "stylelint": "~13.13.1",
+    "stylelint-config-recommended-scss": "~4.3.0",
+    "stylelint-config-standard": "~22.0.0",
+    "stylelint-scss": "~3.20.1",
+    "ts-node": "^10.1.0",
+    "typescript": "~4.3.2"
   },
   "prettier": {
     "singleQuote": true,
