@@ -31,7 +31,7 @@ class MockTranslateService {
 describe('I18nService', () => {
   let i18nService: I18nService;
   let translateService: TranslateService;
-  let onLangChangeSpy: jasmine.Spy;
+  let onLangChangeSpy: jest.Mock;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
@@ -45,12 +45,12 @@ describe('I18nService', () => {
     translateService = TestBed.inject(TranslateService);
 
     // Create spies
-    onLangChangeSpy = jasmine.createSpy('onLangChangeSpy');
-    translateService.onLangChange
-      .subscribe((event: LangChangeEvent) => {
-        onLangChangeSpy(event.lang);
-      });
-    spyOn(translateService, 'use').and.callThrough();
+  onLangChangeSpy = jest.fn();
+  translateService.onLangChange
+    .subscribe((event: LangChangeEvent) => {
+      onLangChangeSpy(event.lang);
+    });
+  jest.spyOn(translateService, 'use');
   });
 
   afterEach(() => {
